Refactor: LottoPrize 패키지 경로 변경
lotto/constant -> lotto/model
