+++ src/test/java/lotto/controller/OutputControllerTest.java
@@ -6,8 +6,8 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import lotto.constant.LottoPrize;
 import lotto.model.Lotto;
+import lotto.model.LottoPrize;
 import lotto.view.OutputView;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
