+++ src/test/java/lotto/service/PlayerTest.java
@@ -5,11 +5,11 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import lotto.constant.LottoPrize;
 import lotto.model.Lotto;
 import lotto.model.LottoGenerator;
 import lotto.model.LottoMachine;
 import lotto.model.LottoNotRandomGenerator;
+import lotto.model.LottoPrize;
 import lotto.model.LottoShop;
 import lotto.model.Player;
 import lotto.model.WinningLotto;
